requirements.txt
langchain==0.3.18
langchain-anthropic==0.3.0
langchain-chroma==0.1.4
langchain-community==0.0.38
langchain-core==0.3.35
langchain-groq==0.2.4
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.1.147
lazy_loader==0.4
groq==0.18.0
python-dotenv==1.0.1
streamlit==1.42.0
streamlit-chat==0.1.1
streamlit-option-menu==0.4.0
sympy==1.13.1